requirements.txt
requests
python-dotenv
pandas
numpy
matplotlib
seaborn
sentence-transformers
faiss-cpu
tqdm
scikit-learn
google-genai
